pyproject.toml
[tool.poetry]
name = "Property friends real estate case"
version = "0.1.0"
description = "Estimate property valuations"
authors = ["Iván González <ivan.gonzalez.prz@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pandas = "^2.1.4"
scikit-learn = "^1.3.2"
category-encoders = "^2.6.3"
jupyter = "^1.0.0"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.26.0"}
cookiecutter = "^2.5.0"
pydantic = "^2.0.0"
requests = "^2.22.0"
loguru = "^0.7.0"
joblib = "^1.2.0"
python-dotenv = "^1.0.0"
luigi = {extras = ["toml"], version = "^3.5.0"}
mlflow = "^2.9.2"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.25"
tabulate = "^0.9.0"
types-tabulate = "^0.9.0.20240106"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
isort = "^5.13.2"
mypy = "^1.8.0"
flake8 = "^7.0.0"
bandit = "^1.7.6"
pytest = "^7.2"
autopep8 = "^2.0.0"
ipdb = "^0.13.0"
pylint = "^3.0.0"
types-pyyaml = "^6.0.12.12"
diagrams = "^0.23.4"


[tool.poetry.group.aws]
optional = true

[tool.poetry.group.aws.dependencies]
mangum = "^0.17.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
  \.toml
  |\.sh
  |\.git
  |\.ini
  |Dockerfile
  |Jenkinfile
)/
'''
